+++ jax/experimental/pallas/gpu.py
@@ -12,11 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Contains Triton specific Pallas functions."""
-from jax._src.pallas import triton
+"""Triton-specific Pallas APIs."""
+
 from jax._src.pallas.triton import approx_tanh
 from jax._src.pallas.triton import elementwise_inline_asm
-
-get_compute_capability = triton.get_compute_capability
-
-del triton