+++ tests/pallas/export_back_compat_pallas_test.py
@@ -26,7 +26,6 @@ from jax._src import test_util as jtu
 from jax._src.internal_test_util import export_back_compat_test_util as bctu
 from jax._src.internal_test_util.export_back_compat_test_data.pallas import cuda_add_one
 from jax.experimental import pallas as pl
-from jax.experimental.pallas import gpu as plgpu
 
 config.parse_flags_with_absl()
 
@@ -40,7 +39,7 @@ class CompatTest(bctu.CompatTestBase):
     if not jtu.test_device_matches(["gpu"]):
       self.skipTest("Only works on GPU")
     if (jtu.test_device_matches(["cuda"]) and
-        plgpu.get_compute_capability(0) < 80):
+        not jtu.is_device_gpu_at_least("8.0")):
       self.skipTest("Only works on GPUs with capability >= sm80")
     super().setUp()