+++ tests/pallas/gpu_attention_test.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 
 import os
-import unittest
 
 from absl.testing import absltest
 from absl.testing import parameterized
@@ -28,10 +27,6 @@ import numpy as np
 
 os.environ["XLA_PYTHON_CLIENT_MEM_FRACTION"] = "0.5"
 
-try:
-  from jax.experimental.pallas import gpu as plgpu
-except ImportError:
-  pass
 # pylint: disable=no-value-for-parameter
 
 
@@ -39,21 +34,12 @@ config.update("jax_traceback_filtering", "off")
 config.parse_flags_with_absl()
 
 
-class PallasTest(jtu.JaxTestCase):
-
-  def check_gpu_capability_at_least(self, capability, device: int = 0):
-    return plgpu.get_compute_capability(device) >= capability
+class DecodeAttentionTest(jtu.JaxTestCase):
 
   def setUp(self):
-    if not jtu.test_device_matches(["gpu"]):
-      self.skipTest("Only works on GPU")
-    try:
-      import triton  # noqa: F401
-    except ImportError:
-      self.skipTest("Triton is not installed. Skipping PallasTest.")
     super().setUp()
-
-class DecodeAttentionTest(PallasTest):
+    if not jtu.is_device_gpu_at_least("8.0"):
+      self.skipTest("Fused attention only works on GPUs with capability >= sm80")
 
   @parameterized.named_parameters(*[
       (
@@ -86,10 +72,6 @@ class DecodeAttentionTest(PallasTest):
       kwargs,
   ):
     del kwargs
-    if not self.check_gpu_capability_at_least(80):
-      raise unittest.SkipTest(
-          "Fused attention only works on GPUs with capability >= sm80"
-      )
 
     k1, k2, k3 = random.split(random.key(0), 3)
     q = random.normal(k1, (batch_size, num_heads, head_dim), dtype=jnp.float16)
@@ -134,10 +116,6 @@ class DecodeAttentionTest(PallasTest):
       kwargs,
   ):
     del kwargs
-    if not self.check_gpu_capability_at_least(80):
-      raise unittest.SkipTest(
-          "Fused attention only works on GPUs with capability >= sm80"
-      )
 
     k1, k2, k3 = random.split(random.key(0), 3)
     q = random.normal(